release: Add build-connect target to Makefile
Add a `build-connect` target to the `Makefile` to build Teleport Connect
via yarn on MacOS. Linux uses the `build.assets/Makefile` `teleterm`
target, and Windows build pipelines do not use `make`.

Add the `CSC_NAME` make variable containing the Developer Key ID to tell
electron-builder which key to use to sign the package it produces.

This gives us a little more control over how the Connect built and
packaged and will simplify the CI scripts to have them just call make.
It is also required in order to set the `CSC_NAME` environment name
correctly as the developer ID is determined by the Makefile.